--- capstone/.travis.yml
@@ -1,13 +1,0 @@
-language: node_js
-node_js:
-- stable
-cache:
-  directories:
-  - node_modules
-script:
-- npm run lint
-- npm test
-- npm run build
-- npm run lint-functions
- notifications:
-    slack: blackturtleneck:EW3d3bAeZV1dX5JISpUrQT19
