--- capstone/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "capstone",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "google-maps-react": "^1.1.2",
-    "react": "^16.2.0",
-    "react-available-times": "^1.2.0",
-    "react-dom": "^16.2.0",
-    "react-geolocation": "^1.0.3",
-    "react-responsive-carousel": "^3.1.33",
-    "react-select": "^1.2.1",
-    "react-slick": "^0.19.0"
-  },
-  "devDependencies": {
-    "firebase": "^4.9.1",
-    "react-edit-inline": "^1.0.8",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "1.1.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  }
-}
