--- capstone/src/Enums.js
@@ -1,26 +1,0 @@
-exports.PageContent = {
-  MESSENGER: "messenger",
-  PROFILE: "profile",
-  EDIT_PROFILE: "edit",
-  //   temp
-  DATE_SELECTION: "date selection"
-};
-
-exports.Routes = {
-  MESSENGER: "messenger",
-  PROFILE: "profile",
-  EDIT_PROFILE: "edit",
-  DATE_SELECTION: "date-selection"
-};
-
-exports.Icons = {
-  ANIMALS: "fas fa-paw",
-  TRAVEL: "fas fa-plane",
-  FOOD: "fas fa-utensils",
-  MUSIC: "fas fa-music",
-  SPORTS: "fas fa-futbol",
-  MOVIES: "fas fa-film",
-  TECH: "fas fa-mobile",
-  GAMING: "fas fa-gamepad",
-  NATURE: "fas fa-tree"
-};
