--- capstone/src/UserList.css
@@ -1,21 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Lato");
-
-.user-list-wrapper {
-  width: 20vw;
-  display: inline-block;
-  font-family: "Lato", sans-serif;
-  vertical-align: top;
-  border-right: 1px lightgray solid;
-  height: 90vh;
-  bottom: 0;
-  position: absolute;
-}
-
-.user-list-component {
-  padding: 30px;
-  padding-left: 40px;
-  padding-bottom: 10px;
-  width: 20vw;
-  text-align: left;
-  font-size: 15px;
-}
